backend/requirements.txt
# requirements.txt

Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
flask_jwt_extended
flask_cors
alembic
psycopg2-binary
python-dotenv
gunicorn
Flask-CORS
trueskill
marshmallow
Flask-Marshmallow
marshmallow-enum